Migrate tests from Espresso/androidTest to Robolectric/test (JVM-based)
Replaces the emulator-dependent connectedAndroidTest pipeline with
Robolectric unit tests that run on the JVM — no emulator needed.
Fixes the CI integration-tests failure caused by the emulator runner.

- Moves all test classes to src/test/ with RobolectricTestRunner
- Seeds ShadowPackageManager so app lists are non-empty without a device
- Replaces BfbVisibleIdlingResource with flushBackgroundThreadScheduler
- Updates CI workflow to use ./gradlew testDebugUnitTest
