+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ on:
     branches: [ main ]
 
 jobs:
-  integration-tests:
+  unit-tests:
     runs-on: ubuntu-latest
 
     steps:
@@ -26,24 +26,12 @@ jobs:
           echo "acra_username=" > secret.properties
           echo "acra_password=" >> secret.properties
 
-      - name: Enable KVM for hardware acceleration
-        run: |
-          echo 'KERNEL=="kvm", GROUP="kvm", MODE="0666", OPTIONS+="static_node=kvm"' | sudo tee /etc/udev/rules.d/99-kvm4all.rules
-          sudo udevadm control --reload-rules
-          sudo udevadm trigger --name-match=kvm
-
-      - name: Run integration tests on Android emulator
-        uses: reactivecircus/android-emulator-runner@v2
-        with:
-          api-level: 29
-          target: default
-          arch: x86_64
-          profile: pixel_2
-          script: ./gradlew connectedAndroidTest
+      - name: Run unit tests
+        run: ./gradlew testDebugUnitTest
 
       - name: Upload test results
         if: always()
         uses: actions/upload-artifact@v4
         with:
-          name: integration-test-results
-          path: app/build/reports/androidTests/connected/
+          name: unit-test-results
+          path: app/build/reports/tests/testDebugUnitTest/
